test: e2e test for "return request list" (GH-5691)
